#31 refactor: 코드 정리 및 matchingState->playerMatchingState로 리팩토링
