--- src/store/atoms/matchingState.ts
@@ -1,9 +1,0 @@
-import { atom } from 'recoil';
-
-export const matchingState = atom({
-  key: 'matching',
-  default: {
-    time: '',
-    introduction: '',
-  },
-});
