+++ src/store/index.ts
@@ -1,1 +1,2 @@
-export {};
+export { default as matchingStepState } from './atoms/matchingStepState';
+export { default as playerMatchingState } from './atoms/playerMatchingState';
